move nginx to nginx_metrics
