+++ src/sources/mod.rs
@@ -32,7 +32,7 @@ pub mod logplex;
 #[cfg(feature = "sources-mongodb_metrics")]
 pub mod mongodb_metrics;
 #[cfg(feature = "sources-nginx_metrics")]
-pub mod nginx;
+pub mod nginx_metrics;
 #[cfg(feature = "sources-prometheus")]
 pub mod prometheus;
 #[cfg(feature = "sources-socket")]
